# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  service.ReportSummary:
    properties:
      adr:
        type: number
      occupancy:
        type: number
      occupancy_by_date:
        additionalProperties:
          format: float64
          type: number
        type: object
      revenue:
        type: number
      revpar:
        type: number
      total_bookings:
        type: integer
    type: object
info:
  contact: {}
  description: REST API for hotel booking management (guest, booking, admin inventory,
    reports).
  title: Hotel Booking API
  version: "1.0"
paths:
...
  /admin/reports/summary:
    get:
      parameters:
      - description: Property ID
        in: query
        name: property_id
        type: string
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start
        required: true
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.ReportSummary'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get summary report
      tags:
      - Reports
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"